Put service implementation into its own package
This takes the implementation of the multitentant service (the message
bus, databases, etc.) and puts it under `service/`.

This involves an awful lot of upheaval, which can be summarised as:

 - anything to do with instances (including the type `InstanceID`) is
   under service/ now, and not imported from outside there and
   `cmd/fluxsvc`.

 - the commands in `fluxctl` are reduced to those in the common
   API (api/api.go); getting and setting config, in particular, is no
   longer something you do with the CLI.

 - in fact there's very little config anyway, since most of it is
   replaced by arguments or files (mounted from secrets)

A couple of things I have effectively disabled for now, like GitHub
deploy key integration, because they will need reimplementing in
slightly different forms anyway.